update to use encode instead
