+++ .gitignore
@@ -8,3 +8,4 @@ pkg
 *.lock
 .idea
 .dev/
+.bundle/
